if user installing/uninstalling is root, do not use sudo command

+++ uninstall.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
 cd /usr/local/bin
-sudo rm cls
+
+if [ $(whoami) == "root" ]
+    then 
+        sudo rm cls
+    else 
+        rm cls 
+fi
 
 echo -e "\ncls is no longer installed."
